+++ source/standalone_examples/api/isaacsim.sensors.experimental.physics/contact_sensor.py
@@ -26,7 +26,7 @@ import carb
 import isaacsim.core.experimental.utils.app as app_utils
 import isaacsim.core.experimental.utils.stage as stage_utils
 import numpy as np
-from isaacsim.core.experimental.objects import GroundPlane
+from isaacsim.core.experimental.objects import DistantLight, GroundPlane
 from isaacsim.core.experimental.prims import Articulation
 from isaacsim.core.simulation_manager import SimulationManager
 from isaacsim.sensors.experimental.physics import Contact, ContactSensor
@@ -44,6 +44,7 @@ if assets_root_path is None:
 
 stage_utils.set_stage_units(meters_per_unit=1.0)
 GroundPlane("/World/GroundPlane")
+DistantLight("/World/DistantLight").set_intensities(1000)
 asset_path = assets_root_path + "/Isaac/Robots/IsaacSim/Ant/ant.usd"
 stage_utils.add_reference_to_stage(usd_path=asset_path, path="/World/Ant")